requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-dotenv==1.0.1
selenium==4.16.0
webdriver-manager==4.0.1
pytesseract==0.3.10
Pillow==10.2.0
